pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "cooklang"
authors = [{ name = "Alex Guinman", email = "alex@guinman.id.au" }]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.9",
    "Operating System :: OS Independent",
]
keywords = ["geojson", "KML"]
requires-python = ">=3.8"
dynamic = ["version", "description"]
dependencies = ["typer", "python-frontmatter", "python-slugify"]

[project.optional-dependencies]
test = ["pytest >=2.7.3", "pytest-cov", "mypy"]
dev = ["black", "flake8", "flake8-bugbear", "pep8-naming", "flake8-builtins"]

[project.urls]
Source = "https://github.com/aguinane/Cooklang"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "-ra --failed-first --showlocals --durations=3 --cov=cooklang"

[tool.coverage.run]
omit = ["*/version.py"]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80
